move decoding functionalities to separate module
