+++ rust/catalyst-voting/src/vote_protocol/voter/mod.rs
@@ -1,10 +1,9 @@
 //! Module containing all primitives related to the voter.
 
+mod decoding;
 pub mod proof;
 
-use std::io::Read;
-
-use anyhow::{anyhow, ensure};
+use anyhow::ensure;
 use rand_core::CryptoRngCore;
 
 use crate::crypto::{
@@ -40,40 +39,6 @@ impl EncryptionRandomness {
 }
 
 impl EncryptedVote {
-    /// Decode `EncryptedVote` from bytes.
-    ///
-    /// # Errors
-    ///   - Cannot decode ciphertext.
-    pub fn from_bytes(mut bytes: &[u8], size: usize) -> anyhow::Result<Self> {
-        let mut ciph_buf = [0u8; Ciphertext::BYTES_SIZE];
-
-        let ciphertexts = (0..size)
-            .map(|i| {
-                bytes.read_exact(&mut ciph_buf)?;
-                Ciphertext::from_bytes(&ciph_buf)
-                    .map_err(|e| anyhow!("Cannot decode ciphertext at {i}, error: {e}"))
-            })
-            .collect::<anyhow::Result<_>>()?;
-
-        Ok(Self(ciphertexts))
-    }
-
-    /// Get a deserialized bytes size
-    #[must_use]
-    pub fn bytes_size(&self) -> usize {
-        self.0.len() * Ciphertext::BYTES_SIZE
-    }
-
-    /// Encode `EncryptedVote` tos bytes.
-    #[must_use]
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut res = Vec::with_capacity(self.bytes_size());
-        self.0
-            .iter()
-            .for_each(|c| res.extend_from_slice(&c.to_bytes()));
-        res
-    }
-
     /// Get the ciphertext to the corresponding `voting_option`.
     pub(crate) fn get_ciphertext_for_choice(&self, voting_option: usize) -> Option<&Ciphertext> {
         self.0.get(voting_option)
@@ -132,21 +97,7 @@ pub fn encrypt_vote<R: CryptoRngCore>(
 
 #[cfg(test)]
 mod tests {
-    use proptest::sample::size_range;
-    use test_strategy::proptest;
-
     use super::*;
-
-    #[proptest]
-    fn encrypted_vote_to_bytes_from_bytes_test(
-        #[any(size_range(0..u8::MAX as usize).lift())] ciphers: Vec<Ciphertext>,
-    ) {
-        let vote1 = EncryptedVote(ciphers);
-        let bytes = vote1.to_bytes();
-        assert_eq!(bytes.len(), vote1.bytes_size());
-        let vote2 = EncryptedVote::from_bytes(&bytes, vote1.0.len()).unwrap();
-        assert_eq!(vote1, vote2);
-    }
 
     #[test]
     fn vote_test() {
